refactor(long)!: assign liquidation penalty to tick upon first position insertion, and use this value for later calculations
BREAKING CHANGE: `getLiquidationPenalty`, `setLiquidationPenalty` and the `LiquidationPenaltyUpdated` event now all use `uint8` for the value. The `getLongPosition` function now returns a second parameter with the liquidation penalty value for the tick containing this position. On creation of a new long position, the liquidation penalty for that position will not necessarily be the value obtained by `getLiquidationPenalty`, but might be the value stored in the tick that will hold that position.

+++ src/interfaces/UsdnProtocol/IUsdnProtocolEvents.sol
@@ -221,7 +221,7 @@ interface IUsdnProtocolEvents {
      * @notice Emitted when the liquidationPenalty is updated.
      * @param newLiquidationPenalty The new liquidationPenalty.
      */
-    event LiquidationPenaltyUpdated(uint24 newLiquidationPenalty);
+    event LiquidationPenaltyUpdated(uint8 newLiquidationPenalty);
 
     /**
      * @notice Emitted when the safetyMargin is updated.
